chore: Reorganized custom assertions and improved test reporter
