+++ test/lib/custom-assertions/Readme.md
@@ -1,0 +1,3 @@
+This module is a collection of helper functions to assert expectations
+in tests. Each file represents a single assertion function, and _must_ export
+a named function.
